requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-community
openai
python-dotenv
uvloop
httptools
redis
